Process the package outside the controller
